Fix nn.Module for deep inheritance chains.
The nn.Module class makes use of a trick to wrap `dataclasses` initializers in
order to guarantee that .name and .parent attributes appear _after_ attributes
defined in the module itself. This works well for one-level of inheritance.

But when inheritance occurs in more than a single level, as shown in the change
in module_test.py, we see the following error:

```
TypeError: non-default argument 'baz' follows default argument
```

Because for the grand-child class, the dataclasses wrapper does not work as
intended, and the .name and .parent attributes end up showing up in one of its
base classes.

This CL addresses the above problem by making sure the entire class hierarchy is
properly wrapped, not only the immediate parent class.

PiperOrigin-RevId: 369244920